refactor environment enums into types file
